+++ include/CQChartsImagePlot.h
@@ -19,6 +19,7 @@ class CQChartsImagePlotType : public CQChartsPlotType {
   QString name() const override { return "image"; }
   QString desc() const override { return "Image"; }
 
+  // no dimension (uses whole model)
   Dimension dimension() const override { return Dimension::NONE; }
 
   void addParameters() override;
@@ -37,7 +38,7 @@ class CQChartsImagePlotType : public CQChartsPlotType {
 class CQChartsImagePlot;
 
 /*!
- * \brief Image Plot object
+ * \brief Image Plot cell object
  */
 class CQChartsImageObj : public CQChartsPlotObj {
   Q_OBJECT
@@ -63,42 +64,73 @@ class CQChartsImageObj : public CQChartsPlotObj {
   double yColorValue(bool relative) const override;
 
  private:
-  const CQChartsImagePlot* plot_  { nullptr };
-  int                      row_   { -1 };
-  int                      col_   { -1 };
-  double                   value_ { 0.0 };
-  QModelIndex              ind_;
+  const CQChartsImagePlot* plot_  { nullptr }; //!< parent plot
+  int                      row_   { -1 };      //<! row
+  int                      col_   { -1 };      //<! column
+  double                   value_ { 0.0 };     //<! value
+  QModelIndex              ind_;               //<! model index
 };
 
 //---
+
+CQCHARTS_NAMED_SHAPE_DATA(Cell,cell)
+CQCHARTS_NAMED_TEXT_DATA(CellLabel,cellLabel)
+CQCHARTS_NAMED_TEXT_DATA(XLabel,xLabel)
+CQCHARTS_NAMED_TEXT_DATA(YLabel,yLabel)
 
 /*!
  * \brief Image Plot
  */
 class CQChartsImagePlot : public CQChartsPlot,
- public CQChartsObjTextData<CQChartsImagePlot> {
+ public CQChartsObjCellShapeData    <CQChartsImagePlot>,
+ public CQChartsObjCellLabelTextData<CQChartsImagePlot>,
+ public CQChartsObjXLabelTextData   <CQChartsImagePlot>,
+ public CQChartsObjYLabelTextData   <CQChartsImagePlot> {
   Q_OBJECT
 
-  Q_PROPERTY(double minValue        READ minValue          WRITE setMinValue       )
-  Q_PROPERTY(double maxValue        READ maxValue          WRITE setMaxValue       )
-  Q_PROPERTY(bool   xLabels         READ isXLabels         WRITE setXLabels        )
-  Q_PROPERTY(bool   yLabels         READ isYLabels         WRITE setYLabels        )
-  Q_PROPERTY(bool   cellLabels      READ isCellLabels      WRITE setCellLabels     )
-  Q_PROPERTY(bool   scaleCellLabels READ isScaleCellLabels WRITE setScaleCellLabels)
-  Q_PROPERTY(bool   balloon         READ isBalloon         WRITE setBalloon        )
+  // value range
+  Q_PROPERTY(double minValue READ minValue WRITE setMinValue)
+  Q_PROPERTY(double maxValue READ maxValue WRITE setMaxValue)
 
-  CQCHARTS_TEXT_DATA_PROPERTIES
+  // cell
+  Q_PROPERTY(CellStyle cellStyle READ cellStyle WRITE setCellStyle)
+
+  Q_PROPERTY(bool cellLabels READ isCellLabels WRITE setCellLabels)
+
+//Q_PROPERTY(bool scaleCellLabels READ isScaleCellLabels WRITE setScaleCellLabels)
+
+  CQCHARTS_NAMED_SHAPE_DATA_PROPERTIES(Cell,cell)
+
+  CQCHARTS_NAMED_TEXT_DATA_PROPERTIES(CellLabel,cellLabel)
+
+  // x/y labels
+  Q_PROPERTY(bool xLabels READ isXLabels WRITE setXLabels)
+
+  CQCHARTS_NAMED_TEXT_DATA_PROPERTIES(XLabel,xLabel)
+
+  Q_PROPERTY(bool yLabels READ isYLabels WRITE setYLabels)
+
+  CQCHARTS_NAMED_TEXT_DATA_PROPERTIES(YLabel,yLabel)
+
+  Q_ENUMS(CellStyle);
 
  public:
+  enum class CellStyle {
+    RECT,
+    BALLOON
+  };
+
   CQChartsImagePlot(CQChartsView *view, const ModelP &model);
 
   //---
 
-  int numRows() const { return nr_; }
+  // dimension
+  int numRows   () const { return nr_; }
   int numColumns() const { return nc_; }
 
   //---
 
+  // value range
   double minValue() const { return minValue_; }
   void setMinValue(double r);
 
@@ -107,18 +139,25 @@ class CQChartsImagePlot : public CQChartsPlot,
 
   //---
 
+  // cell labels
+  bool isCellLabels     () const { return cellLabels_; }
+//bool isScaleCellLabels() const { return scaleCellLabels_; }
+
+  // x labels
   bool isXLabels() const { return xLabels_; }
+
+  // y labels
   bool isYLabels() const { return yLabels_; }
-
-  bool isCellLabels() const { return cellLabels_; }
-
-  bool isScaleCellLabels() const { return scaleCellLabels_; }
 
   //---
 
-  bool isBalloon() const { return balloon_; }
-  void setBalloon(bool b);
+  // cell style
+  const CellStyle &cellStyle() const { return cellStyle_; }
 
+  bool isRectStyle   () const { return cellStyle_ == CellStyle::RECT; }
+  bool isBalloonStyle() const { return cellStyle_ == CellStyle::BALLOON; }
+
+  // balloon min/max size
   double minBalloonSize() const { return minBalloonSize_; }
   void setMinBalloonSize(double r) { minBalloonSize_ = r; }
 
@@ -135,6 +174,10 @@ class CQChartsImagePlot : public CQChartsPlot,
 
   //---
 
+  bool probe(ProbeData &probeData) const;
+
+  //---
+
   bool addMenuItems(QMenu *menu) override;
 
   //---
@@ -148,12 +191,17 @@ class CQChartsImagePlot : public CQChartsPlot,
   CQChartsGeom::BBox annotationBBox() const override;
 
  public slots:
-  void setXLabels(bool b);
-  void setYLabels(bool b);
+  void setRectStyle   (bool b);
+  void setBalloonStyle(bool b);
+
+  void setCellStyle(const CellStyle &style);
 
   void setCellLabels(bool b);
 
-  void setScaleCellLabels(bool b);
+  void setXLabels(bool b);
+  void setYLabels(bool b);
+
+//void setScaleCellLabels(bool b);
 
  private:
   void addImageObj(int row, int col, double x, double y, double dx, double dy,
@@ -163,17 +211,17 @@ class CQChartsImagePlot : public CQChartsPlot,
   void drawYLabels(QPainter *) const;
 
  private:
-  bool   xLabels_         { false }; //!< x labels
-  bool   yLabels_         { false }; //!< y labels
-  bool   cellLabels_      { false }; //!< cell labels
-  bool   scaleCellLabels_ { false }; //!< scale cell labels
-  bool   balloon_         { false }; //!< draw balloon
-  int    nc_              { 0 };     //!< number of grid columns
-  int    nr_              { 0 };     //!< number of grid rows
-  double minValue_        { 0.0 };   //!< min value
-  double maxValue_        { 0.0 };   //!< max value
-  double minBalloonSize_  { 0.1 };   //!< min balloon size (cell fraction)
-  double maxBalloonSize_  { 1.0 };   //!< max balloon size (cell fraction)
+  CellStyle cellStyle_       { CellStyle::RECT }; //!< cell style
+  bool      cellLabels_      { false };           //!< cell labels
+//bool      scaleCellLabels_ { false };           //!< scale cell labels
+  bool      xLabels_         { false };           //!< x labels
+  bool      yLabels_         { false };           //!< y labels
+  int       nc_              { 0 };               //!< number of grid columns
+  int       nr_              { 0 };               //!< number of grid rows
+  double    minValue_        { 0.0 };             //!< min value
+  double    maxValue_        { 0.0 };             //!< max value
+  double    minBalloonSize_  { 0.1 };             //!< min balloon size (cell fraction)
+  double    maxBalloonSize_  { 1.0 };             //!< max balloon size (cell fraction)
 };
 
 #endif